refactor seeder
