+++ prisma/schema.prisma
@@ -19,7 +19,7 @@ model User {
   first_name  String?
   middle_name String?
   last_name   String?
-  student_id  String  @unique
+  student_id  String? @unique
   role        Role?   @relation(fields: [role_id], references: [id])
   role_id     String?
 
@@ -39,9 +39,9 @@ model User {
 
 model Credential {
   id         String  @id @default(cuid())
-  student_id String  @unique
+  student_id String? @unique
   email      String  @unique
-  password   String?
+  // password   String?
   user       User?
 
   access_token  String   @db.LongText
